pyproject.toml
[project]
name = "daily-briefing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyzotero>=1.7.6",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "flask>=3.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.25.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "html2text>=2024.2.26",
    "google-genai>=0.2.0",
    "psycopg2-binary>=2.9.9",
    "tenacity>=8.2.3",
    "cryptography>=42.0.0",
    "pydantic>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
zotero-digest = "src.cli.main:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]

[tool.pytest.ini_options]
# Filter out deprecation warnings from third-party dependencies
filterwarnings = [
    # Suppress httplib2 deprecation warnings (from google-auth-httplib2 dependency)
    "ignore::DeprecationWarning:httplib2",
    # Suppress pyparsing deprecation warnings (used by httplib2)
    "ignore::DeprecationWarning:pyparsing",
]
